requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
easyocr==1.7.1
filelock==3.12.4
fire==0.5.0
fsspec==2023.10.0
fugashi==1.3.0
huggingface-hub==0.17.3
idna==3.4
imageio==2.31.6
jaconv==0.3.4
Jinja2==3.1.2
lazy_loader==0.3
loguru==0.7.2
manga-ocr==0.1.11
MarkupSafe==2.1.3
mpmath==1.3.0
mss==9.0.1
networkx==3.2
ninja==1.11.1.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.0.1
pyclipper==1.3.0.post5
pyperclip==1.8.2
python-bidi==0.4.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-image==0.22.0
scipy==1.11.3
shapely==2.0.2
six==1.16.0
sympy==1.12
termcolor==2.3.0
tifffile==2023.9.26
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.1
triton==2.1.0
typing_extensions==4.8.0
unidic-lite==1.0.8
urllib3==2.0.7
